% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_connect_server.R
\name{calculate_taxes_http}
\alias{calculate_taxes_http}
\title{Calculate taxes using https}
\usage{
calculate_taxes_http(to_taxsim_tmp_filename, idtl)
}
\arguments{
\item{to_taxsim_tmp_filename}{Filename for temporary file to send data to for uploading.}

\item{idtl}{0 if returning a subset of columns, 2 if returning all columns.}
}
\description{
Calculate taxes using https
}
\keyword{internal}
