% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_connect_server.R
\name{calculate_taxes_ssh}
\alias{calculate_taxes_ssh}
\title{Use SSH to calculate taxes through NBER's server}
\usage{
calculate_taxes_ssh(.data, port = 22)
}
\arguments{
\item{.data}{Dataset that can be sent to NBER's server}

\item{port}{Port number to use for SSH connection (defaults to 22)}
}
\description{
Use SSH to calculate taxes through NBER's server
}
\keyword{internal}
