% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_connect_server.R
\name{calculate_taxes_wasm}
\alias{calculate_taxes_wasm}
\title{Use WASM to calculate taxes locally}
\usage{
calculate_taxes_wasm(.data)
}
\arguments{
\item{.data}{Dataset that can be sent to WASM.}
}
\description{
Use WASM to calculate taxes locally
}
\keyword{internal}
