% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_check_data.R
\name{check_data}
\alias{check_data}
\title{Ensure input data set has required fields and datatypes are correct}
\usage{
check_data(.data, cols, state_column_name)
}
\arguments{
\item{.data}{A data frame containing the input parameters for the TAXSIM 32 program. The column names of the input parameters are below. The column can be in any order.}

\item{cols}{The column names, as a string, in the data set \code{.data}}

\item{state_column_name}{The column name of the state column.}
}
\value{
The function does not return a value. It either produces a stop message with the error or prints that all checks were successful.
}
\description{
Check to ensure all the required column are present and data types are correct. This function binds all the checks through helper functions.
}
\keyword{internal}
