% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_check_data.R
\name{check_dependents}
\alias{check_dependents}
\title{Ensure if there are dependent ages then there are also dependents}
\usage{
check_dependents(.data)
}
\arguments{
\item{.data}{A data frame containing the input parameters for the TAXSIM 32 program. The column names of the input parameters are below. The column can be in any order.}
}
\description{
Ensure if there are dependent ages then there are also dependents
}
\keyword{internal}
