% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_check_data.R
\name{check_filing_status}
\alias{check_filing_status}
\title{Ensure values for filing status 'mstat' are proper.}
\usage{
check_filing_status(filing_status_vector)
}
\arguments{
\item{filing_status_vector}{Column, as a vector, containing filing status}
}
\description{
Ensure values for filing status 'mstat' are proper.
}
\keyword{internal}
