% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_check_data.R
\name{check_greater_zero}
\alias{check_greater_zero}
\title{Check that columns are greater than zero}
\usage{
check_greater_zero(.data, cols)
}
\arguments{
\item{.data}{A data frame containing the input parameters for the TAXSIM 32 program. The column names of the input parameters are below. The column can be in any order.}

\item{cols}{The column names, as a string, in the data set \code{.data}.}
}
\description{
Some columns must have all values greater than zero. Check to make sure this is true.
}
\keyword{internal}
