% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_check_data.R
\name{check_id_number}
\alias{check_id_number}
\title{Check that the \code{id_number} column is an integer and every value is unique.}
\usage{
check_id_number(id_number_col)
}
\arguments{
\item{id_number_col}{Vector that id the \code{id_number} column. This will always be the column \code{id_number} in the input data frame.}
}
\description{
The \code{id_number} column requires a data type of integer and unique value. Check to make sure this is true..
}
