% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_check_data.R
\name{check_numeric}
\alias{check_numeric}
\title{Check numeric columns}
\usage{
check_numeric(.data, cols)
}
\arguments{
\item{.data}{A data frame containing the input parameters for the TAXSIM 35 program. The column names of the input parameters are below. The column can be in any order.}

\item{cols}{The column names, as a string, in the data set \code{.data}.}
}
\description{
Checks that each column which should be numeric or integer is numeric or integer.
}
\keyword{internal}
