% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_check_data.R
\name{check_required_cols}
\alias{check_required_cols}
\title{Ensure the required columns are present}
\usage{
check_required_cols(cols)
}
\arguments{
\item{cols}{The column names, as a string, in the data set \code{.data}}
}
\description{
Ensure the required columns are present
}
\keyword{internal}
