% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_check_data.R
\name{check_spouse}
\alias{check_spouse}
\title{Ensure single taxpayers do not have spouse ages or income}
\usage{
check_spouse(.data, cols)
}
\arguments{
\item{.data}{A data frame containing the input parameters for the TAXSIM 32 program. The column names of the input parameters are below. The column can be in any order.}

\item{cols}{The column names, as a string, in the data set \code{.data}.}
}
\description{
Ensure single taxpayers do not have spouse ages or income
}
\keyword{internal}
