% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_check_data.R
\name{check_state}
\alias{check_state}
\title{Check state column}
\usage{
check_state(.data, cols, state_column_name)
}
\arguments{
\item{.data}{A data frame containing the input parameters for the TAXSIM 32 program. The column names of the input parameters are below. The column can be in any order.}

\item{cols}{The column names, as a string, in the data set \code{.data}.}

\item{state_column_name}{The column name of the state column.}
}
\description{
State should be either a two letter abbreviation or full state name. Check to make sure this is true.
}
