% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_check_data.R
\name{check_taxsimid}
\alias{check_taxsimid}
\title{Check that the \code{taxsimid} column is an integer and every value is unique.}
\usage{
check_taxsimid(taxsimid_col)
}
\arguments{
\item{taxsimid_col}{Vector that id the \code{taxsimid} column. This will always be the column \code{taxsimid} in the input data frame.}
}
\description{
The \code{taxsimid} column requires a whole number and unique value. Check to make sure this is true.
}
\keyword{internal}
