% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_clean_data.R
\name{clean_from_taxsim}
\alias{clean_from_taxsim}
\title{Clean final TAXSIM data set.}
\usage{
clean_from_taxsim(from_taxsim)
}
\arguments{
\item{from_taxsim}{The data set received from TAXSIM.}
}
\value{
Data frame containing the row's \code{id_number} and tax calculations. This data frame can be
merged with the original input data frame by \code{id_number}.
}
\description{
Clean the data set received from TAXSIM by renaming columns and removing columns not needed in
the final output.
}
\keyword{internal}
