% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_connect_server.R
\name{connect_server_all_ssh}
\alias{connect_server_all_ssh}
\title{Upload and download data from all ports}
\usage{
connect_server_all_ssh(to_taxsim_tmp_filename, from_taxsim_tmp_filename)
}
\arguments{
\item{to_taxsim_tmp_filename}{Full file path and name to the temp file containing the data to upload to TAXSIM.}

\item{from_taxsim_tmp_filename}{Full file path and name to the temp file that will contain the downloaded data.}
}
\description{
Tries to upload and download the data from the TAXSIM server using three different ports.
Will use the first port that is successful and the other ports will not be tried.
}
\keyword{internal}
