% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_connect_server.R
\name{connect_server_single}
\alias{connect_server_single}
\title{Upload and download data}
\usage{
connect_server_single(to_taxsim_tmp_filename, from_taxsim_tmp_filename, port)
}
\arguments{
\item{to_taxsim_tmp_filename}{Full file path and name to the temp file containing the data to upload to TAXSIM.}

\item{from_taxsim_tmp_filename}{Full file path and name to the temp file that will contain the downloaded data.}

\item{port}{String. The port to use when connecting to the TAXAIM server}
}
\description{
Uploads the data to the TAXSIM server via SSH and then downloads the data with taxes.
The uploads and downloads are of csv files that are saved in temp directories.
}
\keyword{internal}
