% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_connect_server.R
\name{connect_server_single_ssh}
\alias{connect_server_single_ssh}
\title{Upload and download data}
\usage{
connect_server_single_ssh(
  to_taxsim_tmp_filename,
  from_taxsim_tmp_filename,
  std_error_filename,
  known_hosts_file
)
}
\arguments{
\item{to_taxsim_tmp_filename}{Full file path and name to the temp file containing the data to upload to TAXSIM.
This is stdin.}

\item{from_taxsim_tmp_filename}{Full file path and name to the temp file that will contain the downloaded data.
This is stdout.}

\item{std_error_filename}{The file name for the file to write out errors. This is stderror.}

\item{known_hosts_file}{File name to write out the known_hosts file. Must be in a temp directory
because that is the only directory where you can edit files.}
}
\description{
Uploads the data to the TAXSIM server via SSH and then downloads the data with taxes.
The uploads and downloads are of csv files that are saved in temp directories. The parameters are the
same as those in 'create_ssh_command'.
}
\keyword{internal}
