% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_clean_data.R
\name{convert_marginal_tax_rates}
\alias{convert_marginal_tax_rates}
\title{Recode marginal tax rates.}
\usage{
convert_marginal_tax_rates(marginal_tax_rate_specification)
}
\arguments{
\item{marginal_tax_rate}{String representing the \code{marginal_tax_rate} parameter in \code{taxsim_calculate_taxes}}
}
\description{
Marginal tax rates are specified with the \code{marginal_tax_rates} parameter. The possible values are
descriptive strings. But,TAXSIM requires integers. Convert descriptive strings to integers.
}
\keyword{internal}
