% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_clean_data.R
\name{convert_na}
\alias{convert_na}
\title{Convert NA values to either 0 or the proper state value}
\usage{
convert_na(.data, cols_to_convert)
}
\description{
Convert NA values to either 0 or the proper state value
}
\keyword{internal}
