% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_taxes.R
\name{create_dataset_for_taxsim}
\alias{create_dataset_for_taxsim}
\title{Convert a data frame to the TAXSIM 35 output.}
\usage{
create_dataset_for_taxsim(.data)
}
\arguments{
\item{.data}{The data set used to calculate taxes from}
}
\value{
A data frame that that can be manually uploaded to \href{http://taxsim.nber.org/taxsim35/}{TAXSIM 35}.
}
\description{
This function takes a data set that is in the format required for \code{\link{taxsim_calculate_taxes}},
checks it to make sure it is in the proper format for TAXSIM 35, and then cleans so it can be sent to TAXSIM 35.
This function is useful for troubleshooting. It is not needed to calculate taxes. The function is useful
if you continue receiving unreasonable errors from \code{\link{taxsim_calculate_taxes}}. In such as case,
you can run this function on the data frame used in \code{\link{taxsim_calculate_taxes}}. You should then save
this data frame as a csv file. Then, upload the file to \href{http://taxsim.nber.org/taxsim35/}{TAXSIM 35}.
If there are no errors with TAXSIM 35 then the issue lies in \code{\link{taxsim_calculate_taxes}}.
}
\examples{

family_income <- data.frame(
    id_number = c(1, 2),
     state = c('North Carolina', 'NY'),
     tax_year = c(2015, 2015),
     filing_status = c('single', 'married, jointly'),
     primary_wages = c(10000, 100000),
     primary_age = c(26, 36)
)

family_taxes <- create_dataset_for_taxsim(family_income)

\dontrun{
# write out the data frame as a csv file for uploading to TAXSIM 35
vroom::vroom_write(family_taxes, delim = ",")
}

}
