% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_connect_server.R
\name{create_ssh_command}
\alias{create_ssh_command}
\title{Create SSH command}
\usage{
create_ssh_command(to_taxsim_tmp_filename, from_taxsim_tmp_filename, port)
}
\arguments{
\item{to_taxsim_tmp_filename}{Full file path and name to the temp file containing the data to upload to TAXSIM.}

\item{from_taxsim_tmp_filename}{Full file path and name to the temp file that will contain the downloaded data.}

\item{port}{String. The port to use when connecting to the TAXAIM server}
}
\value{
Returns a string that represents the ssh command.
}
\description{
Create SSH command
}
\keyword{internal}
