% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_connect_server.R
\name{execute_ssh_command}
\alias{execute_ssh_command}
\title{Execute SSH command in a cross-platform way}
\usage{
execute_ssh_command(command)
}
\arguments{
\item{command}{The SSH command to execute}
}
\value{
The command output or an error
}
\description{
Execute SSH command in a cross-platform way
}
\keyword{internal}
