% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_clean_data.R
\name{from_taxsim_cols}
\alias{from_taxsim_cols}
\title{Map output column names.}
\usage{
from_taxsim_cols()
}
\description{
Map the output column names required in this package to the input column names required by TAXSIM.
}
\keyword{internal}
