% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_connect_server.R
\name{import_data_ftp}
\alias{import_data_ftp}
\title{Upload and import TAXSIM results via ftp}
\usage{
import_data_ftp(to_taxsim_tmp_filename, idtl)
}
\arguments{
\item{to_taxsim_tmp_filename}{Full file path and name to the temp file containing the data to upload to TAXSIM.}
}
\value{
A Dataframe of TAXSIM results.
}
\description{
Upload and import TAXSIM results via ftp
}
\keyword{internal}
