% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_connect_server.R
\name{import_data_helper}
\alias{import_data_helper}
\title{Import data helper}
\usage{
import_data_helper(raw_data, idtl)
}
\arguments{
\item{raw_data}{The raw data from TAXSIM that we want to load into R.}
}
\description{
When returning all columns, the data retrieved from TAXSIM contains an extra column. This function
trims the extra column if we are returning all information.
}
\keyword{internal}
