% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_connect_server.R
\name{import_data_ssh}
\alias{import_data_ssh}
\title{Import data via ssh}
\usage{
import_data_ssh(
  to_taxsim_tmp_filename,
  from_taxsim_tmp_filename,
  std_error_filename,
  known_hosts_file,
  idtl
)
}
\arguments{
\item{to_taxsim_tmp_filename}{Full file path and name to the temp file containing the data to upload to TAXSIM.
This is stdin.}

\item{from_taxsim_tmp_filename}{Full file path and name to the temp file that will contain the downloaded data.
This is stdout.}

\item{std_error_filename}{The file name for the file to write out errors. This is stderror.}

\item{known_hosts_file}{File name to write out the known_hosts file. Must be in a temp directory
because that is the only directory where you can edit files.}

\item{idtl}{Whether all columns are being returned. 0 if only returning the primary columns.
2 if returning all columns from TAXSIM}
}
\description{
This function wraps all the other ssh functions to upload, download, and import TAXSIM results.
It's the only function that need sto be ran to import data via ssh
}
\keyword{internal}
