% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_connect_server.R
\name{is_ssh_available}
\alias{is_ssh_available}
\title{Check if SSH is available on the system}
\usage{
is_ssh_available()
}
\value{
Logical indicating if SSH is available
}
\description{
Check if SSH is available on the system
}
\keyword{internal}
