% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_clean_data.R
\name{recode_filing_status}
\alias{recode_filing_status}
\title{Recode filing status.}
\usage{
recode_filing_status(filing_status_colname)
}
\arguments{
\item{filing_status_colname}{Column, as a vector, containing filing status}
}
\value{
Vector with integers reflecting numeric value of filing status.
}
\description{
Check to make sure the strings in \code{filing_status} are correct and recode from a string to an integer.
}
\keyword{internal}
