% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{taxpayer_finances}
\alias{taxpayer_finances}
\title{Financial and household characteristics of 1,000 taxpayer units.}
\format{
A data frame with 1,000 rows and 16 variables. Variable definitions can be
found in the following article: \url{https://www.shaneorr.io/r/usincometaxes/articles/taxsim-input.html}
}
\source{
Created through random data generation.
}
\usage{
taxpayer_finances
}
\description{
A data set containing financial and household characteristics of 1,000 taxpayer units.
The data set was randomly generated and does not reflect real data.
It is formatted and ready for use in the \code{usincometaxes} package.
}
\keyword{datasets}
