% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_clean_data.R
\name{taxsim_cols}
\alias{taxsim_cols}
\title{Map input column names.}
\usage{
taxsim_cols()
}
\description{
Map the input column names required in this package to the input column names required by TAXSIM.
}
\keyword{internal}
