% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usincometaxes.R
\docType{package}
\name{usincometaxes}
\alias{usincometaxes}
\title{usincometaxes: A package for calculating state and federal income taxes in the United States.}
\description{
The \code{usincometaxes} package is a single function packages that calculates state and federal income
taxes in the United State. It relies on the NBER's TAXSIM 32 tax calculator for the calculations.
See \url{https://users.nber.org/~taxsim/taxsim32/} for more information on TAXSIM 32.
}
