% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_connect_server.R
\name{calculate_taxes_http}
\alias{calculate_taxes_http}
\title{Calculate taxes using https}
\usage{
calculate_taxes_http(.data, to_taxsim_tmp_filename)
}
\arguments{
\item{.data}{Dataset to send to NBER via http.}

\item{to_taxsim_tmp_filename}{Filename for temporary file to send data to for uploading.}
}
\description{
Calculate taxes using https
}
\keyword{internal}
