% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_check_data.R
\name{check_parameters}
\alias{check_parameters}
\title{Check input parameters}
\usage{
check_parameters(.data, marginal_tax_rates, return_all_information, interface)
}
\description{
Check that the input parameters to \code{taxsim_calculate_taxes} are of the proper type
The parameters to this function should be the same as those to \code{taxsim_calculate_taxes}
}
\keyword{internal}
