% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_taxes.R
\name{create_dataset_for_taxsim}
\alias{create_dataset_for_taxsim}
\title{Create a data set to send to TAXSIM 32.}
\usage{
create_dataset_for_taxsim(.data)
}
\description{
This function takes a data set, checks it to make sure it is in the proper format for TAXSIM 32,
and then cleans so it can be sent to TAXSIM 32.
}
\details{
None
}
