% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_connect_server.R
\name{create_ssh_command}
\alias{create_ssh_command}
\title{Create SSH command}
\usage{
create_ssh_command(
  to_taxsim_tmp_filename,
  from_taxsim_tmp_filename,
  std_error_filename,
  known_hosts_file,
  port
)
}
\arguments{
\item{to_taxsim_tmp_filename}{Full file path and name to the temp file containing the data to upload to TAXSIM.
This is stdin.}

\item{from_taxsim_tmp_filename}{Full file path and name to the temp file that will contain the downloaded data.
This is stdout.}

\item{std_error_filename}{The file name for the file to write out errors. This is stderror.}

\item{known_hosts_file}{File name to write out the known_hosts file. Must be in a temp directory
because that is the only directory where you can edit files.}

\item{port}{String. The port to use when connecting to the TAXSIM server}
}
\value{
Returns a string that represents the ssh command.
}
\description{
Create SSH command
}
\keyword{internal}
