% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_clean_data.R
\name{get_state_soi}
\alias{get_state_soi}
\title{Get state SOI from name.}
\usage{
get_state_soi(state_column)
}
\arguments{
\item{state_column}{Vectors containing the states to calculate taxes for. Generally, this is the
state column from the data set that will be sent to TAXSIM.}
}
\value{
Named integer vector with each number between 1 and 51 representing the state's SOI.
Names are the state's two letter abbreviation.
}
\description{
Return the integer number state SOI of a state based on either its two letter abbreviation or
full name.
}
\keyword{internal}
