% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_connect_server.R
\name{import_data_ssh}
\alias{import_data_ssh}
\title{Import data via ssh}
\usage{
import_data_ssh(to_taxsim_tmp_filename, from_taxsim_tmp_filename, idtl)
}
\arguments{
\item{to_taxsim_tmp_filename}{Full file path and name to the temp file containing the data to upload to TAXSIM.}

\item{from_taxsim_tmp_filename}{Full file path and name to the temp file that will contain the downloaded data.}
}
\description{
This function wraps all the other ssh functions to upload, download, and import TAXSIM results.
It's the only function that need sto be ran to import data via ssh
}
\keyword{internal}
