% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_taxes.R
\name{taxsim_calculate_taxes}
\alias{taxsim_calculate_taxes}
\title{Calculate state and federal taxes using TAXSIM 32.}
\usage{
taxsim_calculate_taxes(
  .data,
  return_all_information = FALSE,
  upload_method = "ftp"
)
}
\arguments{
\item{.data}{Data frame containing the information that will be used to calculate taxes.
This data set will be sent to TAXSIM. Data frame must have specified column names and data types.}

\item{return_all_information}{Boolean (TRUE or FALSE). Whether to return all information from TAXSIM (TRUE),
or only key information (FALSE). Returning all information returns 42 columns of output, while only
returning key information returns 9 columns. It is faster to download results with only key information.}

\item{upload_method}{Either 'ftp' or 'ssh', can also use upper case. Defaults to 'ftp'. Determines whether ftp or ssh will be used to send data
to TAXSIM and retrieve the results. SSH is faster, so use it when there are over 100,000 records.
SSH is available in Windows 10 since autumn of 2019.
If you run an earlier Windows, try the built-in ftp client instead.}
}
\value{
Returns a data frame with the following columns:

\code{id_number} The unique id number for the row that corresponds to the id number in \code{.data}

\code{federal_taxes} Total federal taxes due. Corresponds to \code{fiitax} in TAXSIM.

\code{state_taxes} Total state taxes due; if state taxes were calculated.
Corresponds to \code{siitax} in TAXSIM.

\code{fica_taxes} Total FICA taxes due. This includes both the employee and company share.
Corresponds to \code{fica} in TAXSIM.

\code{federal_marginal_rate} Federal marginal tax rate of taxpayer. Corresponds to \code{frate} in TAXSIM.

\code{state_marginal_rate} State marginal tax rate of taxpayer. Corresponds to \code{srate} in TAXSIM.

\code{fica_rate} FICA rate. Corresponds to \code{ficar} in TAXSIM.
}
\description{
This function calculates state and federal income taxes using the TAXSIM 32 tax simulator.
See \url{https://users.nber.org/~taxsim/taxsim32/} for more information on TAXSIM 32.
}
\section{Required columns}{


\code{id_number} An arbitrary, non-negative, integer. This number links the results from TAXSIM
32 to the original data frame with entries.

\code{tax_year} Tax year ending Dec 31 (4 digits between 1960 and 2023). State must be zero if
year is before 1977 or after 2023.

\code{filing_status} Filing status of tax unit. One of the following: "single" for single;
"married, jointly" for married, filing jointly; "married, separately" for married, filing separately;
"dependent child" for dependent, usually a child with income; or "head of household" for head of household filing status.
}

\section{Optional columns}{


\code{state} State two letter abbreviation or state SOI code. If state income taxes are not needed,
either label as "No State" or remove this variable. State income tax information is only available from 1977 to 2023.

\code{primary_age} Age of primary taxpayer as of December 31st of tax year. Taxpayer age variables
determine eligibility for additional standard deductions, personal exemption, EITC and AMT exclusion.

\code{spouse_age} Age of spouse as of December 31st of tax year.

\code{num_dependents} Total number of dependents.

\code{num_dependents_thirteen} Number of children under 13 with eligible child care expenses (Dependent Care Credit).

\code{num_dependents_seventeen} Number of children under 17 for the entire tax year (Child Credit).
This includes children under 13.

\code{num_dependents_eitc} Number of qualifying children for EITC. (Typically younger than 19 or
younger than 24 and a full-time student).

\code{primary_wages} Wage and salary income of Primary Taxpayer (include self-employment but no QBI).

\code{spouse_wages} Wage and salary income of Spouse (include self-employment but no QBI). Must
be zero or the column should not exist for non-joint returns.

\code{dividends} Dividend income (qualified dividends only for 2003 on).

\code{interest} Interest income received (+/-).

\code{short_term_capital_gains} Short Term Capital Gains or losses (+/-).

\code{long_term_capital_gains} Long Term Capital Gains or losses (+/-).

\code{other_property_income} Other property income subject to NIIT, including: unearned or limited
partnership and passive S-Corp profits; rent not eligible for QBI deduction; non-qualified
dividends; capital gains distributions on form 1040; and other income or loss not otherwise enumerated here.

\code{other_non_property_income} Other non-property income not subject to Medicare NIIT such as:
alimony; nonwage fellowships; and state income tax refunds (itemizers only). Also includes
adjustments and items such as: alimony paid; Keogh and IRA contributions; foreign income exclusion; and NOLs

\code{pensions} Taxable Pensions and IRA distributions.

\code{social_security} Gross Social Security Benefits.

\code{unemployment} Unemployment compensation received.

\code{other_transfer_income} Other non-taxable transfer income such as: welfare; workers comp;
veterans benefits; and child support that would affect eligibility for state property tax
rebates but would not be taxable at the federal level.

\code{rent_paid} Rent paid (used only for calculating state property tax rebates).

\code{property_taxes} Real Estate taxes paid. This is a preference for the AMT and is is also
used to calculate state property tax rebates.

\code{other_itemized_deductions} Other Itemized deductions that are a preference for the Alternative
Minimum Tax. These would include: Other state and local taxes (line 8 of Schedule A) plus
local income tax; Preference share of medical expenses; Miscellaneous (line 27).

\code{child_care_expenses} Child care expenses.

\code{misc_deductions} Deductions not included in \code{other_itemized_deductions} and not a preference
for the AMT, including (on Schedule A for 2009). Might include: Deductible medical expenses
not included in Line 16; Motor Vehicle Taxes paid; Home mortgage interest; Charitable contributions;
and Casulty or Theft Losses.

\code{scorp_income} Active S-Corp income (is SSTB).

\code{qualified_business_income} Primary Taxpayer's Qualified Business Income (QBI) subject to a
preferential rate without phaseout and assuming sufficient wages paid or capital to be eligible
for the full deduction. Subject to SECA and Medicare additional Earnings Tax.

\code{specialized_service_trade} Primary Taxpayer's Specialized Service Trade or Business service
(SSTB) with a preferential rate subject to claw-back. Subject to SECA and Medicare Additional Earnings Tax.

\code{spouse_qualified_business_income} Spouse's QBI. Must be zero for non-joint returns, or the
column should not exist.

\code{spouse_specialized_service_trade} Spouse's SSTB. Must be zero for non-joint returns, or the
column should not exist.
}

\section{Note on number of dependents}{


\code{num_dependents} columns are not mutually exclusive. For example, a family with a 13 year old
can report the dependent in \code{num_dependents_thirteen} and also in \code{num_dependents_seventeen}.
}

\section{Additional Output}{
NA
}

\section{Giving credit where it is due}{


The NBER's \href{http://users.nber.org/~taxsim/taxsim32/}{TAXSIM 32} tax simulator does all tax
calculations. This package simply lets users interact with the tax simulator through R. Therefore,
users should cite the TAXSIM 32 tax simulator when they use this package in their work:

Feenberg, Daniel Richard, and Elizabeth Coutts, An Introduction to the TAXSIM Model,
Journal of Policy Analysis and Management vol 12 no 1, Winter 1993, pages 189-194.
}

\examples{

family_income <- data.frame(
    id_number = as.integer(c(1, 2)),
     state = c('North Carolina', 'NY'),
     tax_year = c(2015, 2015),
     filing_status = c('single', 'married, jointly'),
     primary_wages = c(10000, 100000),
     primary_age = c(26, 36)
)

family_taxes <- taxsim_calculate_taxes(family_income, upload_method = 'ftp')

merge(family_income, family_taxes, by = 'id_number')

}
